Retry if we use a bad page token.
LDAP cookies don't persist across connections, so if we resume a sync with an invalid page token, we error out. This fixes the issue by starting the ldap search from scratch. It retries exactly one time and fails if we get the error again, so it shouldn't cause issues with infinite loops for permission errors.